feat: 바코드 인식 반납 API 코드 작성

+++ backend/src/main/java/com/pororo/docar/domain/book/service/BookApiResponse.java
@@ -52,7 +52,7 @@ public class BookApiResponse {
             @NoArgsConstructor
             @AllArgsConstructor
             public static class Packing {
-                private double sizeDepth;
+                private Long sizeDepth;
 
                 // 게터/세터 메서드
             }
